requirements.txt
Pillow==10.4.0
rich==13.3.2
pick==2.2.0
